Interim commit
It compiles, but still having the worst time trying to figure out why
ScalarEvaluator doesn't chase that sign-extension up to the load and see
that the array index is "I" (and thus is affine).

This also contains some scratch code where I tried to write down what we
need to do, independent of this LLVM crap.