Start working on our OWN ScalarEvolution process
I don't know if it is my misunderstanding or if ScalarEvolution is not
fully implemented, but it is damn sure not serving our needs.

The general gist of what we need is to chase an array subscript access
and find out if it is in the form aX+b (well, technically it could be
aX+bY+C and so on for as many loop indices as we have, but stay with
me). The only way it could be aX+b is if there is a tree (in the AST
tree sense) containing only one (optional) multiply and one (optional)
add and a load instruction that loads in the loop index.

This commit does not yet deal with the multiply, but it does the rest of
what is described [for one index variable only].